fix(mne): clear the MSVC /W4 warnings and two bugs they exposed
With fiff clean the Windows leg advanced into src/libraries/mne and reported 39
warnings there. Two of them were real defects, not style:

MNEInverseOperator::check_ch_names took a parameter named `info`, which hid the
member of the same name, so `data_ch_names = info.ch_names` read the operator's
own channel list instead of the caller's. The compatibility check MinimumNorm
performs before applying an operator to evoked data could therefore never fail.
Renaming the parameter turned it into an unused-parameter error, which is what
surfaced the bug.

read_vol_geom parsed the FreeSurfer volume-geometry header with sscanf("%s")
into `char eq[2]`, an unbounded write into a two-byte buffer that any source
space file could overflow. Replaced the whole loop with QByteArray tokens, which
also retires the nine C4996 deprecation warnings in that function.

The rest are mechanical: float literals for the C4305 truncations, explicit
casts for the size_t narrowings, and renamed locals for the remaining shadowing
reports. mne_sensor_locations moves from FILE* to QTextStream for its C4996.

Verified: clang build 0 warnings, g++ sweep over 1410 units reports no
diagnostics, 180/180 ctest including the MNE-Python cross-validation.

+++ src/libraries/mne/mne_sss_data.cpp
@@ -151,7 +151,7 @@ std::unique_ptr<MNESssData> MNESssData::read_from_node(QSharedPointer<FiffStream
             return nullptr;
         }
         {
-            int ncomp = t_pTag->size()/sizeof(fiff_int_t);
+            int ncomp = static_cast<int>(t_pTag->size()/sizeof(fiff_int_t));
             int *raw  = t_pTag->toInt();
             s->comp_info = Eigen::VectorXi::Map(raw, ncomp);
